requirements.txt
exchangelib
google_auth_oauthlib
google-api-python-client
google-auth-httplib2
dateutils